// ui/src/api/types.ts
export interface components {
    schemas: {
        /** @enum {string} */
        "ServerResponseStatus.SUCCESS": "success";
        Player: {
            id: string;
            name: string;
        };
        Game: {
            id: string;
            players: components["schemas"]["Player"][];
        };
        /** @enum {string} */
        ServerResponseStatus: "success" | "error";
        ServerSuccess_Game_: {
            status: components["schemas"]["ServerResponseStatus.SUCCESS"];
            data: components["schemas"]["Game"];
        };
        /** @enum {string} */
        "ServerResponseStatus.ERROR": "error";
        ServerError: {
            status: components["schemas"]["ServerResponseStatus.ERROR"];
            error: string;
        };
        GameResponse_Game_: components["schemas"]["ServerSuccess_Game_"] | components["schemas"]["ServerError"];
        JoinGameRequest: {
            playerName: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}